Enhanced Arrays and List in PreviewExpression
replaced with deepToString() which will:
Returns a string representation of the "deep contents" of the specified array.  If the array contains other arrays as elements, the string representation contains their contents and so on.  This method is designed for converting multidimensional arrays to strings.  It avoids infinite recursion, if the array contains itself or a reference to itself which is rendered as "[[...]]"